+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Replace WP browser with codeception for tests
+- Update font awesome social icon class names to v5
 
 ## 0.7.2 - 2018-03-01